add check password when deleted email

+++ backend/api/custome_authen.py
@@ -68,7 +68,7 @@ class AddEmailView(APIView): #Thêm 1 email khác vào tài khoản
             return Response({'Vui lòng nhập email'}, status=400)
         if EmailAddress.objects.filter(email=new_email,verified=True).exists(): # check để tránh khi 1 người nhập đại mà sau này có ng tạo mail y hệt 
             return Response({"error: Email already exist"},status=400)
-        email_address, created = EmailAddress.objects.get_or_create(
+        email_address, created = EmailAddress.objects.get_or_create( # nếu get thì trả false, nếu created thì trả true vì khai báo created trước đó
             user=user,
             email=new_email,
             defaults={
@@ -76,7 +76,9 @@ class AddEmailView(APIView): #Thêm 1 email khác vào tài khoản
                 "verified": False,
             }
         )
-        # nếu get thì trả false, nếu true thì trả true vì khai báo created trước đó
+        if not created:
+            return Response({'error': 'Email already added before'},status=400)
+       
         email_address.send_confirmation(request) #signup false để bảo đây là thêm chứ k phải đăng kí tài khoản mới
         return Response({'Email added successfully, please check your email'},status =200)
         
@@ -95,41 +97,46 @@ class SetPrimaryEmailView(APIView): # đặt 1 email làm mặc đinh
                 status=status.HTTP_400_BAD_REQUEST
             )
         if email_obj.primary:
-            return Response({"detail": "Already primary"})
+            return Response({"detail": "Already primary"},status=400)
         email_obj.set_as_primary()
-        return Response({"detail": "Primary email updated"})
+        return Response({"detail": "Primary email updated"},status=200)
 
 class DeleteEmailView(APIView):
     permission_classes = [IsAuthenticated]
 
     def delete(self, request, pk):
         user = request.user
+        password= request.data.get("password")
+        if not password:
+            return Response({'error':"Please enter password"},status=400)
+        auth = authenticate(username=user.username,password=password)
+        if auth:
+            email_obj = EmailAddress.objects.filter(
+                id=pk,
+                user=user
+            ).first()
 
-        email_obj = EmailAddress.objects.filter(
-            id=pk,
-            user=user
-        ).first()
+            if not email_obj:
+                return Response(
+                    {"error": "Email not found"},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            # Không cho xóa primary
+            if email_obj.primary:
+                return Response(
+                    {"error": "Cannot delete primary email"},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            # Không cho xóa nếu là email duy nhất
+            if EmailAddress.objects.filter(user=user).count() <= 1:
+                return Response(
+                    {"error": "Cannot delete the only email"},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
 
-        if not email_obj:
-            return Response(
-                {"error": "Email not found"},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        # Không cho xóa primary
-        if email_obj.primary:
-            return Response(
-                {"error": "Cannot delete primary email"},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        # Không cho xóa nếu là email duy nhất
-        if EmailAddress.objects.filter(user=user).count() <= 1:
-            return Response(
-                {"error": "Cannot delete the only email"},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        email_obj.delete()
-        return Response({"detail": "Email deleted successfully"})
+            email_obj.delete()
+            return Response({"detail": "Email deleted successfully"},status=200)
+        return Response({'error':'wrong password'},status=400)
     
 class UserEmail(generics.ListAPIView):
     permission_classes=[IsAuthenticated]
@@ -143,11 +150,12 @@ class UserEmail(generics.ListAPIView):
 - thêm signal khi 1 cái verified thì xóa hết 
 '''
 
-class CheckPassword(APIView):
-    permission_classes=[IsAuthenticated]
-    def post(self,request):
-        password=request.data.get("password")
-        user= authenticate(username=request.user.username,password=password)
-        if user:
-            return Response({'valid':'True'})
-        return Response({'valid':'False'})
+# class CheckPassword(APIView):
+#     permission_classes=[IsAuthenticated]
+#     def post(self,request):
+#         password=request.data.get("password")
+#         user= authenticate(username=request.user.username,password=password)
+#         if user:
+#             return Response({'valid':'True'})
+#         return Response({'valid':'False'})
+
